requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==45.0.5
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
PyYAML==6.0.2
scp==0.15.0
tqdm==4.67.1
